bits: add ffs compat, minor test fixes

+++ test/test_common.h
@@ -6,7 +6,7 @@
 
 #define str_check(a, b) tt_str_op(a, ==, b)
 #define int_check(a, b) tt_int_op(a, ==, b)
-#define ull_check(a, b) tt_assert_op_type(a, ==, b, unsigned long long, "%llu")
+#define ull_check(a, b) tt_assert_op_type(a, ==, b, uint64_t, "%" PRIu64)
 
 extern struct testcase_t aatree_tests[];
 extern struct testcase_t cbtree_tests[];
